NEW #33506 Add a constant to apply a ratio to resize images in ODT templates (#33543)
* CLOSE #33506 add a constant holding the ratio to resize images in ODT templates

A new Dolibarr constant 					MAIN_DOC_ODT_IMAGE_RATIO
is taken in consideration for ODT templates.
This constant holds a number that is multiplied when computing the size of images in ODT templates.
All doc_generic_*_odt_modules.php files have been updated to call the setImage Method with a new parameter $ratio
The method setImage in the file odt.php has been modified to have a new parameter $ratio and use it to compute the image size

* FIX #33506 wrong branch

some other files needed to be changed in the develop branch

* FIX #33506 PR errors fixed

* FIX #33506 codesniffer errors fixed

* QUAL #33506 completed method comment