FIX #35011 Computed extrafield on tasks list now display correct values (#35038)
The functions assigns value to `$obj` on line 976, but doesn't assign value to `$object`, that is then assigned to `$objectoffield` in line 75 of `extrafields_list_print_fields.tpl.php`. Also `$objectoffield` is the recommended value to use for computed extrafields, according to the docs.
But assigning `$object` is not enough, because `$objectoffield` at this moment is not grabbed from the global scope, so the value stays local and `dol_eval()` doesn't find it. So it also needs to grab the variable from the global scope to affect it correctly.
The changed function, `projectLinesa()` is called only twice in the project: by itself recursively to go through all subtasks, and on the page where I detected the bug initially, `tasks.php`, so the impact should be minimal.